MLR_and_Supervised_ML_Project - Final
EXECUTIVE SUMMARY: We reported our best two models: MLE with all variables and a lasso regression model.
MLE: We implemented the MLE model with the training and testing dataset and got an RMSE of 0.122 and an adjusted r-square of 0.29, similar to our report 1 model where we got an RMSE of 0.115 and an adjusted r-square of 0.26.

Lasso: Next we implemented our final model, lasso regression. First, we loaded the training and testing datasets, placing a specific seed to avoid random data generation on each model iteration, while preparing “cross validation” as our hyperparameter tuning technique. Like before, we used a built-in library and applied the cross-validation function to tune the lambda. When we ran a testing and training dataset on a previously reported model with 10 folds for cross-validation, we got an RMSE of 0.17 and R-square value of 0.32.

Improvements: For improving our model, we thought of playing around with a number of folds for cross-validation. For 200 k fold cross validation we got RMSE of 0.109 and r square of 0.499.
After an evaluation of a multiple linear regression model and Lasso regression, we concluded that the Lasso regression model with 200 k folds gives us the best results and is the optimal choice for the given dataset.